feat: PUT

+++ src/routes.ts
@@ -20,10 +20,10 @@ export const handleRoutes = async (req: IncomingMessage, res: ServerResponse) =>
         return;
     }
 
-    // if (method === 'PUT' && pathname?.startsWith('/api/users/')) {
-    //     controller.updateUserById(req, res);
-    //     return;
-    // }
+    if (method === 'PUT' && pathname?.startsWith('/api/users/')) {
+        controller.updateUserById(req, res);
+        return;
+    }
 
     if (method === 'DELETE' && pathname?.startsWith('/api/users/')) {
         controller.deleteUserById(req, res);
